fix(axdevice): seal PCI topology construction
