+++ virtualization/axdevice/src/pci/mod.rs
@@ -31,4 +31,5 @@ pub(crate) use runtime::PciBindingLease;
 pub use runtime::{
     EndpointRouteToken, PciBarAccess, PciFunction, PciRootBinding, PciRootBindingKey,
 };
-pub use topology::{PciTopologyBuilder, ResolvedPciBar, ResolvedPciFunction, ResolvedPciTopology};
+pub(crate) use topology::PciTopologyBuilder;
+pub use topology::{ResolvedPciBar, ResolvedPciFunction, ResolvedPciTopology};
