+++ virtualization/axdevice/src/pci/topology.rs
@@ -15,15 +15,15 @@ use crate::{DeviceNodeId, ResourceRequest};
 
 const DEVICE_COUNT: u8 = 32;
 
-/// Mutable PCI topology declaration sealed before VM execution.
-pub struct PciTopologyBuilder {
+/// Internal PCI topology declaration used while resolving a device graph.
+pub(crate) struct PciTopologyBuilder {
     functions: BTreeMap<DeviceNodeId, PciFunctionSpec>,
     reservations: BTreeSet<PciBdf>,
 }
 
 impl PciTopologyBuilder {
     /// Creates an empty topology.
-    pub const fn new() -> Self {
+    pub(crate) const fn new() -> Self {
         Self {
             functions: BTreeMap::new(),
             reservations: BTreeSet::new(),
@@ -36,7 +36,7 @@ impl PciTopologyBuilder {
     ///
     /// Returns [`PciError::DuplicateFunction`] if the stable identity is
     /// already present.
-    pub fn add_function(&mut self, function: PciFunctionSpec) -> PciResult {
+    pub(crate) fn add_function(&mut self, function: PciFunctionSpec) -> PciResult {
         let id = function.id.clone();
         if self.functions.contains_key(&id) {
             return Err(PciError::DuplicateFunction {
@@ -56,27 +56,28 @@ impl PciTopologyBuilder {
     ///
     /// Returns [`PciError::InvalidAddress`] for a segment or bus outside the
     /// currently supported segment-zero, bus-zero topology.
-    pub fn reserve_bdf(&mut self, bdf: PciBdf) -> PciResult {
+    pub(crate) fn reserve_bdf(&mut self, bdf: PciBdf) -> PciResult {
         validate_supported_bdf(bdf)?;
         self.reservations.insert(bdf);
         Ok(())
     }
 
     /// Resolves all fixed requests, performs deterministic automatic
-    /// placement, validates the complete topology, and freezes it.
+    /// placement, and validates the complete topology before graph ownership
+    /// is assigned.
     ///
     /// The resolved functions carry placeholder `owner`/`host` identities
     /// (each pointing at itself) until graph resolution assigns real
-    /// ownership; production callers must go through
-    /// `DeclaredDeviceGraph::resolve`, which always assigns ownership inside
-    /// its transaction before publication.
+    /// ownership. This is an internal intermediate step; callers must use
+    /// [`DeclaredDeviceGraph::resolve`](crate::DeclaredDeviceGraph::resolve)
+    /// for the sealed topology published to firmware and runtime.
     ///
     /// # Errors
     ///
     /// Returns a typed [`PciError`] for malformed host apertures, BDF
     /// conflicts, invalid identities, or BAR placement failures. No partial
     /// topology is returned.
-    pub fn resolve(self, memory_aperture: Range<u64>) -> PciResult<ResolvedPciTopology> {
+    pub(crate) fn resolve(self, memory_aperture: Range<u64>) -> PciResult<ResolvedPciTopology> {
         validate_memory_aperture(&memory_aperture)?;
         let bdfs = resolve_bdfs(&self.functions, &self.reservations)?;
         let bar_addresses = resolve_bar_addresses(&memory_aperture, &self.functions)?;
@@ -338,4 +339,274 @@ fn validate_supported_bdf(bdf: PciBdf) -> PciResult {
         });
     }
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{ConfigOffset, PciClass, PciMemoryBar, PciSegment};
+
+    const APERTURE_START: u64 = 0x2000_0000;
+    const APERTURE_END: u64 = 0x2040_0000;
+    const BAR_SIZE: u64 = 0x1_0000;
+
+    #[test]
+    fn rejects_addresses_outside_conventional_pci_ranges() {
+        assert!(matches!(
+            PciBdf::new(PciSegment::new(0), 0, 32, 0),
+            Err(PciError::InvalidAddress {
+                component: "device",
+                ..
+            })
+        ));
+        assert!(matches!(
+            PciBdf::new(PciSegment::new(0), 0, 0, 8),
+            Err(PciError::InvalidAddress {
+                component: "function",
+                ..
+            })
+        ));
+        assert!(matches!(
+            PciBarIndex::new(6),
+            Err(PciError::InvalidAddress {
+                component: "BAR index",
+                ..
+            })
+        ));
+        assert!(matches!(
+            ConfigOffset::new(0x100),
+            Err(PciError::InvalidAddress {
+                component: "config offset",
+                ..
+            })
+        ));
+    }
+
+    #[test]
+    fn rejects_invalid_memory_bar_sizes() {
+        let bar = PciBarIndex::new(2).unwrap();
+        assert!(matches!(
+            PciMemoryBar::new(bar, 0),
+            Err(PciError::InvalidBar { .. })
+        ));
+        assert!(matches!(
+            PciMemoryBar::new(bar, 0x18),
+            Err(PciError::InvalidBar { .. })
+        ));
+        assert!(matches!(
+            PciMemoryBar::new(bar, 1_u64 << 33),
+            Err(PciError::InvalidBar { .. })
+        ));
+    }
+
+    #[test]
+    fn rejects_absent_vendor_identity_and_invalid_host_apertures() {
+        let mut identity = PciTopologyBuilder::new();
+        identity
+            .add_function(PciFunctionSpec::new(
+                node("absent"),
+                PciEndpointIdentity::new(u16::MAX, 0x5678, PciClass::new(0x05, 0x00, 0x00)),
+            ))
+            .unwrap();
+        assert!(matches!(
+            identity.resolve(APERTURE_START..APERTURE_END),
+            Err(PciError::InvalidEndpointIdentity { .. })
+        ));
+
+        assert!(matches!(
+            PciTopologyBuilder::new().resolve(APERTURE_START..APERTURE_START),
+            Err(PciError::InvalidHostAperture { .. })
+        ));
+        assert!(matches!(
+            PciTopologyBuilder::new().resolve(APERTURE_START..(1_u64 << 32) + 1),
+            Err(PciError::InvalidHostAperture { .. })
+        ));
+    }
+
+    #[test]
+    fn resolves_auto_bdfs_deterministically_and_skips_reservations() {
+        let mut builder = PciTopologyBuilder::new();
+        builder.reserve_bdf(bdf(0, 0)).unwrap();
+        builder.reserve_bdf(bdf(31, 0)).unwrap();
+        builder.add_function(function("beta")).unwrap();
+        builder.add_function(function("alpha")).unwrap();
+
+        let topology = builder.resolve(APERTURE_START..APERTURE_END).unwrap();
+
+        assert_eq!(topology.function(&node("alpha")).unwrap().bdf(), bdf(1, 0));
+        assert_eq!(topology.function(&node("beta")).unwrap().bdf(), bdf(2, 0));
+    }
+
+    #[test]
+    fn rejects_fixed_requests_for_nonzero_functions() {
+        let mut builder = PciTopologyBuilder::new();
+        builder
+            .add_function(function("endpoint").with_bdf(ResourceRequest::Fixed(bdf(3, 1))))
+            .unwrap();
+        assert!(matches!(
+            builder.resolve(APERTURE_START..APERTURE_END),
+            Err(PciError::UnsupportedFunctionPlacement { .. })
+        ));
+    }
+
+    #[test]
+    fn rejects_auto_bar_larger_than_the_aperture() {
+        let bar0 = PciBarIndex::new(0).unwrap();
+        let mut builder = PciTopologyBuilder::new();
+        builder
+            .add_function(
+                function("oversized")
+                    .with_bar(PciMemoryBar::new(bar0, 0x100_0000).unwrap())
+                    .unwrap(),
+            )
+            .unwrap();
+        assert!(matches!(
+            builder.resolve(APERTURE_START..APERTURE_END),
+            Err(PciError::BarApertureExhausted { .. })
+        ));
+    }
+
+    #[test]
+    fn rejects_fixed_bars_at_size_misaligned_addresses() {
+        let bar0 = PciBarIndex::new(0).unwrap();
+        let mut builder = PciTopologyBuilder::new();
+        builder
+            .add_function(
+                function("misaligned")
+                    .with_bar(
+                        PciMemoryBar::new(bar0, BAR_SIZE)
+                            .unwrap()
+                            .with_address(ResourceRequest::Fixed(APERTURE_START + 8)),
+                    )
+                    .unwrap(),
+            )
+            .unwrap();
+        assert!(matches!(
+            builder.resolve(APERTURE_START..APERTURE_END),
+            Err(PciError::InvalidBar { .. })
+        ));
+    }
+
+    #[test]
+    fn rejects_fixed_bdf_conflicts_and_reserved_requests() {
+        let mut duplicate = PciTopologyBuilder::new();
+        duplicate
+            .add_function(function("alpha").with_bdf(ResourceRequest::Fixed(bdf(3, 0))))
+            .unwrap();
+        duplicate
+            .add_function(function("beta").with_bdf(ResourceRequest::Fixed(bdf(3, 0))))
+            .unwrap();
+        assert!(matches!(
+            duplicate.resolve(APERTURE_START..APERTURE_END),
+            Err(PciError::DuplicateBdf { .. })
+        ));
+
+        let mut reserved = PciTopologyBuilder::new();
+        reserved.reserve_bdf(bdf(3, 0)).unwrap();
+        reserved
+            .add_function(function("endpoint").with_bdf(ResourceRequest::Fixed(bdf(3, 0))))
+            .unwrap();
+        assert!(matches!(
+            reserved.resolve(APERTURE_START..APERTURE_END),
+            Err(PciError::BdfReserved { .. })
+        ));
+    }
+
+    #[test]
+    fn places_larger_auto_bars_first_and_preserves_function_order_tiebreaks() {
+        let bar0 = PciBarIndex::new(0).unwrap();
+        let bar2 = PciBarIndex::new(2).unwrap();
+        let mut builder = PciTopologyBuilder::new();
+        builder
+            .add_function(
+                function("beta")
+                    .with_bar(PciMemoryBar::new(bar0, 0x1_0000).unwrap())
+                    .unwrap(),
+            )
+            .unwrap();
+        builder
+            .add_function(
+                function("alpha")
+                    .with_bar(PciMemoryBar::new(bar2, 0x20_0000).unwrap())
+                    .unwrap(),
+            )
+            .unwrap();
+
+        let topology = builder.resolve(APERTURE_START..APERTURE_END).unwrap();
+
+        assert_eq!(
+            topology
+                .function(&node("alpha"))
+                .unwrap()
+                .bar(bar2)
+                .unwrap()
+                .address(),
+            APERTURE_START
+        );
+        assert_eq!(
+            topology
+                .function(&node("beta"))
+                .unwrap()
+                .bar(bar0)
+                .unwrap()
+                .address(),
+            APERTURE_START + 0x20_0000
+        );
+    }
+
+    #[test]
+    fn rejects_fixed_bars_outside_or_overlapping_the_aperture() {
+        let bar0 = PciBarIndex::new(0).unwrap();
+        let mut outside = PciTopologyBuilder::new();
+        outside
+            .add_function(
+                function("outside")
+                    .with_bar(
+                        PciMemoryBar::new(bar0, BAR_SIZE)
+                            .unwrap()
+                            .with_address(ResourceRequest::Fixed(APERTURE_END)),
+                    )
+                    .unwrap(),
+            )
+            .unwrap();
+        assert!(matches!(
+            outside.resolve(APERTURE_START..APERTURE_END),
+            Err(PciError::InvalidBar { .. })
+        ));
+
+        let mut overlap = PciTopologyBuilder::new();
+        for id in ["alpha", "beta"] {
+            overlap
+                .add_function(
+                    function(id)
+                        .with_bar(
+                            PciMemoryBar::new(bar0, BAR_SIZE)
+                                .unwrap()
+                                .with_address(ResourceRequest::Fixed(APERTURE_START)),
+                        )
+                        .unwrap(),
+                )
+                .unwrap();
+        }
+        assert!(matches!(
+            overlap.resolve(APERTURE_START..APERTURE_END),
+            Err(PciError::BarConflict { .. })
+        ));
+    }
+
+    fn function(id: &str) -> PciFunctionSpec {
+        PciFunctionSpec::new(
+            node(id),
+            PciEndpointIdentity::new(0x1234, 0x5678, PciClass::new(0x05, 0x00, 0x00))
+                .with_revision(1),
+        )
+    }
+
+    fn node(id: &str) -> DeviceNodeId {
+        DeviceNodeId::new(id).unwrap()
+    }
+
+    fn bdf(device: u8, function: u8) -> PciBdf {
+        PciBdf::new(PciSegment::new(0), 0, device, function).unwrap()
+    }
 }
